+++ .php-cs-fixer.dist.php
@@ -12,5 +12,8 @@ return (new PhpCsFixer\Config())
         'phpdoc_align' => false,
         'no_superfluous_phpdoc_tags' => false,
         'array_syntax' => ['syntax' => 'short'],
+        'phpdoc_separation' => [
+            'skip_unlisted_annotations' => true,
+        ]
     ])
     ->setFinder($finder);