+++ src/Command/AuthCreateApikeyCommand.php
@@ -3,7 +3,6 @@
 namespace App\Command;
 
 use App\Service\ApiKeyUserService;
-use Exception;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -23,7 +22,7 @@ class AuthCreateApikeyCommand extends Command
     }
 
     /**
-     * @throws Exception
+     * @throws \Exception
      */
     protected function execute(InputInterface $input, OutputInterface $output): int
     {