+++ tests/Api/BookingTest.php
@@ -18,7 +18,6 @@ use App\Service\NotificationServiceInterface;
 use App\Service\WebformService;
 use App\Tests\AbstractBaseApiTestCase;
 use App\Utils\ValidationUtils;
-use Exception;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Security\Core\Security;
@@ -31,7 +30,7 @@ class BookingTest extends AbstractBaseApiTestCase
     use InteractsWithMessenger;
 
     /**
-     * @throws TransportExceptionInterface|Exception
+     * @throws TransportExceptionInterface|\Exception
      */
     public function testBookingVoter(): void
     {
@@ -114,7 +113,7 @@ class BookingTest extends AbstractBaseApiTestCase
     }
 
     /**
-     * @throws Exception
+     * @throws \Exception
      */
     public function testWebformSubmitMessageHandler(): void
     {
@@ -231,7 +230,7 @@ class BookingTest extends AbstractBaseApiTestCase
         $booking->setResourceEmail('test@bookaarhus.local.itkdev.dk');
         $booking->setStartTime(new \DateTime());
         $booking->setEndTime(new \DateTime());
-        $booking->setUserName('auther1');
+        $booking->setUserName('author1');
         $booking->setUserMail('author1@bookaarhus.local.itkdev.dk');
         $booking->setMetaData([
             'meta_data_4' => '1, 2, 3',
@@ -260,6 +259,7 @@ class BookingTest extends AbstractBaseApiTestCase
         $res->setPermissionCitizen(true);
         $res->setPermissionEmployee(true);
         $res->setHasWhitelist(false);
+        $res->setAcceptConflict(false);
 
         $aakResourceRepositoryMock = $this->getMockBuilder(AAKResourceRepository::class)
             ->disableOriginalConstructor()